+++ relief/schema/mappings.py
@@ -47,6 +47,17 @@ class Mapping(Container):
         if new_value is not Unspecified:
             raise AttributeError("can't set attribute")
 
+    def _set_value_from_native(self, value):
+        super(Mapping, self).clear()
+        if value is not Unspecified:
+            if hasattr(self, "_raw_value"):
+                del self._raw_value
+            for key in value:
+                super(Mapping, self).__setitem__(key, _Value(
+                    self.member_schema[0](key),
+                    self.member_schema[1](value[key])
+                ))
+
     def _set_value_from_raw(self, value):
         super(Mapping, self).clear()
         if value is not Unspecified: