+++ tests/schema/conftest.py
@@ -52,6 +52,14 @@ class ElementTest(object):
         assert element.raw_value == possible_value
         assert element.value == possible_value
 
+    def test_set_from_native_invalidates_is_valid(self, element_cls, possible_value):
+        element = element_cls()
+        element.set_from_native(possible_value)
+        assert element.validate()
+        assert element.is_valid
+        element.set_from_native(possible_value)
+        assert element.is_valid is None
+
     def test_set_from_raw_invalidates_is_valid(self, element_cls, possible_value):
         element = element_cls()
         element.set_from_raw(possible_value)
@@ -65,3 +73,13 @@ class ElementTest(object):
         assert a.properties == {'foo': 1}
         b = a.with_properties(bar=2)
         assert b.properties == {'foo': 1, 'bar': 2}
+
+    def test_set_from_native_custom_serialize(self, element_cls, possible_value):
+        class FooElement(element_cls):
+            def serialize(self, value):
+                return 'foo'
+
+        element = FooElement()
+        element.set_from_native(possible_value)
+        assert element.value == possible_value
+        assert element.raw_value == 'foo'