+++ tests/schema/test_core.py
@@ -36,3 +36,11 @@ class TestElement(ElementTest):
         element.set_from_raw(1)
         assert element.value == 1
         assert element.raw_value == 1
+
+    def test_set_from_native(self):
+        element = Element()
+        assert element.value is Unspecified
+        assert element.raw_value is Unspecified
+        element.set_from_raw(1)
+        assert element.value == 1
+        assert element.raw_value == 1